tests/requirements.txt
# Testing framework dependencies
pytest==7.4.0
pytest-cov==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
netifaces==0.11.0
requests==2.31.0
